refactor(veículo): altera inserção de processo de acoplamento para utilizar functions
PL-3210

+++ src/main/java/br/com/zalf/prolog/webservice/frota/veiculo/acoplamento/_model/realizacao/VeiculoAcoplamento.java
@@ -23,7 +23,7 @@ public final class VeiculoAcoplamento {
     @NotNull
     private final VeiculoAcoplamentoAcao acaoRealizada;
     @Range(min = 1, max = 6)
-    private final int posicaoAcaoRealizada;
+    private final short posicaoAcaoRealizada;
     @Nullable
     private final Long kmColetado;
